#71 Correções Codacy Static Code Analysis
