+++ src/main/java/br/com/burnhop/api/resource/PostsResource.java
@@ -1,8 +1,7 @@
 package br.com.burnhop.api.resource;
 
 import br.com.burnhop.api.controller.PostsController;
-import br.com.burnhop.model.Dto.*;
-import br.com.burnhop.model.Posts;
+import br.com.burnhop.model.dto.*;
 import br.com.burnhop.repository.ContentRepository;
 import br.com.burnhop.repository.UsersRepository;
 import br.com.burnhop.repository.PostsRepository;
