+++ src/main/java/br/com/burnhop/model/dto/CreatedLoginDto.java
@@ -1,8 +1,7 @@
-package br.com.burnhop.model.Dto;
+package br.com.burnhop.model.dto;
 
 import br.com.burnhop.model.Login;
 import io.swagger.annotations.ApiModelProperty;
-import io.swagger.annotations.ApiParam;
 
 public class CreatedLoginDto {
