+++ src/main/java/br/com/burnhop/model/dto/GroupDto.java
@@ -1,4 +1,4 @@
-package br.com.burnhop.model.Dto;
+package br.com.burnhop.model.dto;
 
 import br.com.burnhop.model.Groups;
 import io.swagger.annotations.ApiModelProperty;
