+++ src/main/java/br/com/burnhop/model/dto/UpdatedPostDto.java
@@ -1,4 +1,4 @@
-package br.com.burnhop.model.Dto;
+package br.com.burnhop.model.dto;
 
 import io.swagger.annotations.ApiModelProperty;
 
@@ -6,10 +6,6 @@ public class UpdatedPostDto {
 
     @ApiModelProperty(value = "Uma String que representa o conteúdo do post", example = "Conteúdo teste")
     private String content;
-
-    public UpdatedPostDto() {
-
-    }
 
     public String getContent() {
         return content;
