+++ src/main/java/br/com/burnhop/model/dto/UpdatedUserDto.java
@@ -1,14 +1,8 @@
-package br.com.burnhop.model.Dto;
+package br.com.burnhop.model.dto;
 
-import br.com.burnhop.model.Users;
-import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiModelProperty;
-import io.swagger.annotations.ApiParam;
 
-import java.sql.Date;
-import java.sql.Timestamp;
-
-public class CreatedUserDto {
+public class UpdatedUserDto {
 
     @ApiModelProperty(value = "Uma String que representa o nome do usuário", example = "Nome Exemplo Teste")
     private String name;
@@ -19,21 +13,10 @@ public class CreatedUserDto {
     @ApiModelProperty(value = "Uma String que representa data de nascimento do usuário", example = "2020-10-18")
     private String data_nasc;
 
-    private CreatedLoginDto login;
+    private String email;
 
-    public CreatedUserDto() {
+    public UpdatedUserDto() {
 
-    }
-
-    public Users toUser() {
-        Users user = new Users();
-        user.setName(this.name);
-        user.setUsername(this.username);
-        user.setData_nasc(Date.valueOf(this.data_nasc));
-        user.setCreated_on(new Timestamp(System.currentTimeMillis()));
-        user.setLogin(this.login.toLogin());
-
-        return user;
     }
 
     public String getName() {
@@ -60,12 +43,11 @@ public class CreatedUserDto {
         this.data_nasc = data_nasc;
     }
 
-    public CreatedLoginDto getLogin() {
-        return login;
+    public String getEmail() {
+        return email;
     }
 
-    public void setLogin(CreatedLoginDto login) {
-        this.login = login;
+    public void setEmail(String email) {
+        this.email = email;
     }
-
 }
