+++ src/main/java/br/com/burnhop/repository/GroupsRepository.java
@@ -1,9 +1,6 @@
 package br.com.burnhop.repository;
 
 import br.com.burnhop.model.Groups;
-import br.com.burnhop.model.Dto.GroupDto;
-import br.com.burnhop.model.Posts;
-import br.com.burnhop.model.Users;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
