+++ src/test/java/br/com/burnhop/UsersResourceTest.java
@@ -1,46 +1,28 @@
 package br.com.burnhop;
 
-import java.sql.Date;
-import java.sql.Timestamp;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import br.com.burnhop.model.Users;
-import br.com.burnhop.model.Login;
-import br.com.burnhop.model.Dto.CreatedLoginDto;
-import br.com.burnhop.model.Dto.CreatedUserDto;
-import br.com.burnhop.model.Dto.UserDto;
-import br.com.burnhop.api.resource.UsersResource;
+import br.com.burnhop.model.dto.CreatedLoginDto;
+import br.com.burnhop.model.dto.CreatedUserDto;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.http.MediaType;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
 //import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 
-
-import org.springframework.boot.test.mock.mockito.MockBean;
-
-import org.mockito.MockitoAnnotations;
-import org.mockito.Mock;
-
-
-import br.com.burnhop.repository.LoginRepository;
-import br.com.burnhop.repository.UsersRepository;
 
 import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.TestMethodOrder;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
 
-import br.com.burnhop.model.Dto.CreatedPostDto;
+import br.com.burnhop.model.dto.CreatedPostDto;
 
 @SpringBootTest
 @AutoConfigureMockMvc
